@@ -5,14 +5,14 @@ source 'https://rubygems.org'
 ruby '2.7.6'
 
 # CORE DEPS
-gem 'rails', '6.0.5'
-gem 'puma', '>= 4.3.11'
+gem 'rails', '6.1.6'
+gem 'puma', '~> 5.6.4'
 
 # EARLY TO APPLY TO OTHER GEMS
-gem 'dotenv-rails', '~>2.6'
+gem 'dotenv-rails', '~> 2.6'
 
 # DB/MODEL
-gem 'pg', '0.18.4'
+gem 'pg', '1.4.2'
 gem 'ancestry', '~> 3.0.5'
 gem 'ranked-model', '~> 0.4.3'
 gem 'rails_admin', '~> 2.2'
@@ -125,6 +125,7 @@ gem 'webpacker', '~> 3.0.0'
 # ASSET/UI
 gem 'uglifier', require: false
 gem 'kaminari', '~> 1.2.1'
+gem 'sprockets', '~> 3.7.2'
 
 # MIDDLEWARE
 gem 'rack-cache', '~> 1.6.1', require: 'rack/cache'
@@ -140,7 +141,7 @@ gem 'rack-heartbeat'
 gem 'clever-ruby', '~> 2.0.1'
 
 # Memory profiling
-gem 'puma_worker_killer', '~> 0.1.0'
+gem 'puma_worker_killer'
 
 # temp for migrations
 gem 'paperclip'
