@@ -48,7 +48,8 @@ class RuleFeedbackHistory
         entry: f_h.entry,
         highlight: f_h.metadata.instance_of?(Hash) ? f_h.metadata['highlight'] : '',
         session_uid: f_h.feedback_session_uid,
-        strength: f_h.feedback_history_ratings.max_by(&:updated_at)&.rating
+        strength: f_h.feedback_history_ratings.max_by(&:updated_at)&.rating,
+        api: f_h.metadata.instance_of?(Hash) ? f_h.metadata['api'] || {} : {}
     }
   end
 
