@@ -46,7 +46,8 @@ class FeedbackHistory < ApplicationRecord
     AUTO_ML = "autoML",
     SPELLING = "spelling",
     OPINION = "opinion",
-    PREFILTER = "prefilter"
+    PREFILTER = "prefilter",
+    LOW_CONFIDENCE = 'low-confidence'
   ]
   FILTER_TYPES = [
     FILTER_ALL = "all",
