@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-workers Integer(ENV['PUMA_WORKERS'] || 3)
-threads Integer(ENV['MIN_THREADS']  || 1), Integer(ENV['MAX_THREADS'] || 10)
+workers Integer(ENV['WEB_CONCURRENCY'] || 3)
+threads Integer(ENV['MIN_THREADS'] || 1), Integer(ENV['MAX_THREADS'] || 10)
 
 preload_app!
 
