@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Evidence
-  VERSION = "0.0.5"
+  VERSION = "0.0.6"
 end
