@@ -6,7 +6,6 @@ describe ApplicationController, type: :controller do
   let(:user) { create(:user) }
 
   describe "#user_inactive_for_too_long?" do
-
     before do
       ApplicationController.send(:public, *ApplicationController.protected_instance_methods)
     end
@@ -57,5 +56,35 @@ describe ApplicationController, type: :controller do
       end
     end
 
+  end
+
+  context '#handle_invalid_inauthenticity_token' do
+    controller do
+      def custom
+        raise ActionController::InvalidAuthenticityToken
+      end
+    end
+
+    let(:referer) { 'http://test.host/referer_path' }
+    let(:redirect_path) { URI.parse(referer).path }
+
+    before do
+      routes.draw { post 'custom' => "anonymous#custom" }
+      request.headers['HTTP_REFERER'] = referer
+    end
+
+    it 'handles InvalidAuthenticityToken error with format: :json' do
+      post :custom, params: {}, as: :json
+
+      expect(flash[:error]).to eq(I18n.t('actioncontroller.errors.invalid_authenticity_token'))
+      expect(response.body).to eq({ redirect: redirect_path }.to_json)
+    end
+
+    it 'handles InvalidAuthenticityToken error with format: :html' do
+      post :custom, params: {}
+
+      expect(flash[:error]).to eq(I18n.t('actioncontroller.errors.invalid_authenticity_token'))
+      expect(response).to redirect_to(redirect_path)
+    end
   end
 end
